+++ src-tauri/Cargo.toml
@@ -36,9 +36,14 @@ moka = { version = "0.12.7", features = ["sync"] }
 rsntp = { version = "4.0.0", default-features = false, features = ["chrono"]  }
 flate2 = "1.0.30"
 sysinfo = "0.35.1"
+thiserror = "2.0.16"
+strum = "0.27"
+strum_macros = "0.27"
+async-trait = "0.1.89"
 meter-core = { git = "https://github.com/snoww/meter-core-rs", optional = true }
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
+tauri-plugin-shell = "2"
 tauri-plugin-clipboard-manager = "2"
 tauri-plugin-updater = "2.9.0"
 tauri-plugin-opener = "2.4.0"
