+++ src-tauri/src/constants.rs
@@ -6,11 +6,13 @@ use window_vibrancy::Color;
 
 pub const WINDOW_MS: i64 = 5_000;
 pub const WINDOW_S: i64 = 5;
+pub const DB_VERSION: i32 = 5;
 pub const METER_WINDOW_LABEL: &str = "main";
 pub const METER_MINI_WINDOW_LABEL: &str = "mini";
 pub const LOGS_WINDOW_LABEL: &str = "logs";
 pub const DATABASE_PATH: &str = "encounters.db";
 pub const SETTINGS_PATH: &str = "settings.json";
+pub const DEFAULT_SETTINGS_PATH: &str = "settings.template.json";
 pub const LOCAL_PLAYERS_PATH: &str = "local_players.json";
 pub const REGION_PATH: &str = "current_region";
 pub const STEAM_GAME_URL: &str = "steam://rungameid/1599340";
