+++ src-tauri/src/live/status_tracker.rs
@@ -5,7 +5,7 @@ use crate::live::status_tracker::StatusEffectBuffCategory::{BattleItem, Bracelet
 use crate::live::status_tracker::StatusEffectCategory::Debuff;
 use crate::live::status_tracker::StatusEffectShowType::All;
 use crate::live::utils::get_new_id;
-use crate::parser::models::{EncounterEntity, EntityType};
+use crate::models::{EncounterEntity, EntityType};
 use chrono::{DateTime, Duration, Utc};
 use hashbrown::HashMap;
 use meter_core::packets::structures::{PCStruct, StatusEffectData};
@@ -199,7 +199,7 @@ impl StatusTracker {
     ) -> (Vec<StatusEffectDetails>, Vec<StatusEffectDetails>) {
         let timestamp = Utc::now();
 
-        let use_party_for_source = if source_entity.entity_type == EntityType::PLAYER {
+        let use_party_for_source = if source_entity.entity_type == EntityType::Player {
             self.should_use_party_status_effect(source_entity.character_id, local_character_id)
         } else {
             false
@@ -215,7 +215,7 @@ impl StatusTracker {
         let status_effects_on_source =
             self.actually_get_status_effects(source_id, source_type, timestamp);
 
-        let use_party_for_target = if source_entity.entity_type == EntityType::PLAYER {
+        let use_party_for_target = if source_entity.entity_type == EntityType::Player {
             self.should_use_party_status_effect(target_entity.character_id, local_character_id)
         } else {
             false
