+++ src-tauri/src/models/data.rs
@@ -1,0 +1,131 @@
+use crate::models::utils::int_or_string_as_string;
+use hashbrown::HashMap;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct SkillData {
+    pub id: i32,
+    pub name: Option<String>,
+    #[serde(rename = "type", default)]
+    #[serde(deserialize_with = "int_or_string_as_string")]
+    pub skill_type: String,
+    pub cooldown: u32,
+    pub desc: Option<String>,
+    pub class_id: u32,
+    pub icon: Option<String>,
+    pub grade: String,
+    pub identity_category: Option<String>,
+    #[serde(alias = "groups")]
+    pub groups: Option<Vec<i32>>,
+    pub summon_source_skills: Option<Vec<u32>>,
+    pub source_skills: Option<Vec<u32>>,
+    #[serde(default)]
+    pub is_hyper_awakening: bool,
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct SkillEffectData {
+    pub id: i32,
+    pub comment: String,
+    #[serde(skip)]
+    pub stagger: i32,
+    pub source_skills: Option<Vec<u32>>,
+    pub directional_mask: Option<i32>,
+    pub item_name: Option<String>,
+    pub item_desc: Option<String>,
+    pub item_type: Option<String>,
+    pub icon: Option<String>,
+    pub values: Vec<i32>,
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct SkillBuffData {
+    pub id: i32,
+    pub name: Option<String>,
+    pub desc: Option<String>,
+    pub icon: Option<String>,
+    pub icon_show_type: Option<String>,
+    pub duration: i32,
+    // buff | debuff
+    pub category: String,
+    #[serde(rename(deserialize = "type"))]
+    #[serde(deserialize_with = "int_or_string_as_string")]
+    pub buff_type: String,
+    pub status_effect_values: Option<Vec<i32>>,
+    pub buff_category: Option<String>,
+    pub target: String,
+    pub unique_group: u32,
+    #[serde(rename(deserialize = "overlap"))]
+    pub overlap_flag: i32,
+    pub per_level_data: HashMap<String, PerLevelData>,
+    pub source_skills: Option<Vec<u32>>,
+    pub set_name: Option<String>,
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct PerLevelData {
+    pub passive_options: Vec<PassiveOption>,
+    // pub status_effect_values: Vec<i32>
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct PassiveOption {
+    #[serde(rename(deserialize = "type"))]
+    pub option_type: String,
+    pub key_stat: String,
+    pub key_index: i32,
+    pub value: i32,
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+pub struct CombatEffectData {
+    pub effects: Vec<CombatEffectDetail>,
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+pub struct CombatEffectDetail {
+    pub ratio: i32,
+    pub cooldown: i32,
+    pub conditions: Vec<CombatEffectCondition>,
+    pub actions: Vec<CombatEffectAction>,
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+#[serde(rename_all = "camelCase", default)]
+pub struct CombatEffectCondition {
+    #[serde(rename(deserialize = "type"))]
+    pub condition_type: String,
+    pub actor_type: String,
+    pub arg: i32,
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+#[serde(rename_all = "camelCase", default)]
+pub struct CombatEffectAction {
+    pub action_type: String,
+    pub actor_type: String,
+    pub args: Vec<i32>,
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+pub struct Npc {
+    pub id: i32,
+    pub name: Option<String>,
+    pub grade: String,
+    #[serde(rename = "type")]
+    pub npc_type: String,
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+pub struct Esther {
+    pub name: String,
+    pub icon: String,
+    pub skills: Vec<i32>,
+    #[serde(alias = "npcs")]
+    pub npc_ids: Vec<u32>,
+}
