+++ src-tauri/src/models/encounter.rs
@@ -1,0 +1,287 @@
+use hashbrown::{HashMap, HashSet};
+use serde::{Deserialize, Serialize};
+use serde_with::serde_as;
+use crate::models::Skill;
+
+use crate::models::EntityType;
+use crate::models::StatusEffect;
+
+#[derive(Debug, Serialize, Clone, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct Encounter {
+    pub last_combat_packet: i64,
+    pub fight_start: i64,
+    pub local_player: String,
+    pub entities: HashMap<String, EncounterEntity>,
+    pub current_boss_name: String,
+    pub current_boss: Option<EncounterEntity>,
+    pub encounter_damage_stats: EncounterDamageStats,
+    pub duration: i64,
+    pub difficulty: Option<String>,
+    pub favorite: bool,
+    pub cleared: bool,
+    pub boss_only_damage: bool,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub sync: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub region: Option<String>,
+}
+
+#[derive(Debug, Serialize, Clone, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct EncounterDamageStats {
+    pub total_damage_dealt: i64,
+    pub top_damage_dealt: i64,
+    pub total_damage_taken: i64,
+    pub top_damage_taken: i64,
+    pub dps: i64,
+    pub buffs: HashMap<u32, StatusEffect>,
+    pub debuffs: HashMap<u32, StatusEffect>,
+    pub total_shielding: u64,
+    pub total_effective_shielding: u64,
+    pub applied_shield_buffs: HashMap<u32, StatusEffect>,
+    #[serde(skip)]
+    pub unknown_buffs: HashSet<u32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub misc: Option<EncounterMisc>,
+    pub boss_hp_log: HashMap<String, Vec<BossHpLog>>,
+}
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct EncounterEntity {
+    pub id: u64,
+    pub character_id: u64,
+    pub npc_id: u32,
+    pub name: String,
+    pub entity_type: EntityType,
+    pub class_id: u32,
+    pub class: String,
+    pub gear_score: f32,
+    pub current_hp: i64,
+    pub max_hp: i64,
+    pub current_shield: u64,
+    pub is_dead: bool,
+    pub skills: HashMap<u32, Skill>,
+    pub damage_stats: DamageStats,
+    pub skill_stats: SkillStats,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub engraving_data: Option<Vec<String>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub ark_passive_active: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub ark_passive_data: Option<ArkPassiveData>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub spec: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub loadout_hash: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub combat_power: Option<f32>,
+}
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase", default)]
+pub struct ArkPassiveData {
+    pub evolution: Option<Vec<ArkPassiveNode>>,
+    pub enlightenment: Option<Vec<ArkPassiveNode>>,
+    pub leap: Option<Vec<ArkPassiveNode>>,
+}
+
+#[derive(Debug, Default, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase", default)]
+pub struct ArkPassiveNode {
+    pub id: u32,
+    pub lv: u8,
+}
+
+#[serde_as]
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase", default)]
+pub struct DamageStats {
+    pub damage_dealt: i64,
+    pub hyper_awakening_damage: i64,
+    pub damage_taken: i64,
+    pub buffed_by: HashMap<u32, i64>,
+    pub debuffed_by: HashMap<u32, i64>,
+    pub buffed_by_support: i64,
+    pub buffed_by_identity: i64,
+    pub debuffed_by_support: i64,
+    pub buffed_by_hat: i64,
+    pub crit_damage: i64,
+    pub back_attack_damage: i64,
+    pub front_attack_damage: i64,
+    pub shields_given: u64,
+    pub shields_received: u64,
+    pub damage_absorbed: u64,
+    pub damage_absorbed_on_others: u64,
+    pub shields_given_by: HashMap<u32, u64>,
+    pub shields_received_by: HashMap<u32, u64>,
+    pub damage_absorbed_by: HashMap<u32, u64>,
+    pub damage_absorbed_on_others_by: HashMap<u32, u64>,
+    pub deaths: i64,
+    pub death_time: i64,
+    pub dps: i64,
+    #[serde(default)]
+    pub dps_average: Vec<i64>,
+    #[serde(default)]
+    pub dps_rolling_10s_avg: Vec<i64>,
+    pub rdps_damage_received: i64,
+    pub rdps_damage_received_support: i64,
+    pub rdps_damage_given: i64,
+    #[serde(default)]
+    pub incapacitations: Vec<IncapacitatedEvent>,
+}
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct SkillStats {
+    pub casts: i64,
+    pub hits: i64,
+    pub crits: i64,
+    pub back_attacks: i64,
+    pub front_attacks: i64,
+    pub counters: i64,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub identity_stats: Option<String>,
+}
+
+#[derive(Debug)]
+pub struct DamageData {
+    pub skill_id: u32,
+    pub skill_effect_id: u32,
+    pub damage: i64,
+    pub shield_damage: Option<i64>,
+    pub modifier: i32,
+    pub target_current_hp: i64,
+    pub target_max_hp: i64,
+    pub damage_attribute: Option<u8>,
+    pub damage_type: u8,
+}
+
+#[derive(Debug, Serialize, Clone, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct Identity {
+    pub gauge1: u32,
+    pub gauge2: u32,
+    pub gauge3: u32,
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct IncapacitatedEvent {
+    #[serde(rename = "type")]
+    pub event_type: IncapacitationEventType,
+    pub timestamp: i64,
+    pub duration: i64, // in a live meter, this might be retroactively updated to be shortened if the user uses get up or gets incapacitated with the same type again
+}
+
+#[derive(Debug, Serialize, Deserialize, Clone, PartialEq)]
+#[allow(non_camel_case_types)]
+pub enum IncapacitationEventType {
+    FALL_DOWN,
+    CROWD_CONTROL,
+}
+
+pub type IdentityLog = Vec<(i64, (u32, u32, u32))>;
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct IdentityArcanist {
+    // timestamp, (percentage, card, card)
+    pub log: Vec<(i32, (f32, u32, u32))>,
+    pub average: f64,
+    pub card_draws: HashMap<u32, u32>,
+}
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct IdentityArtistBard {
+    // timestamp, (percentage, bubble)
+    pub log: Vec<(i32, (f32, u32))>,
+    pub average: f64,
+}
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct IdentityGeneric {
+    // timestamp, percentage
+    pub log: Vec<(i32, f32)>,
+    pub average: f64,
+}
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase", default)]
+#[serde_as]
+pub struct EncounterMisc {
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub boss_hp_log: Option<HashMap<String, Vec<BossHpLog>>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub raid_clear: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub party_info: Option<HashMap<i32, Vec<String>>>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub region: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub version: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub rdps_valid: Option<bool>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub rdps_message: Option<String>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub ntp_fight_start: Option<i64>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub manual_save: Option<bool>,
+}
+
+#[derive(Serialize, Deserialize, Debug, Clone, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct BossHpLog {
+    pub time: i32,
+    pub hp: i64,
+    #[serde(default)]
+    pub p: f32,
+}
+
+impl BossHpLog {
+    pub fn new(time: i32, hp: i64, p: f32) -> Self {
+        Self { time, hp, p }
+    }
+}
+
+#[derive(Debug, Default, Deserialize, Clone)]
+pub struct EngravingData {
+    pub id: u32,
+    pub name: Option<String>,
+    pub icon: Option<String>,
+}
+
+#[derive(Debug, PartialEq)]
+#[allow(non_camel_case_types)]
+#[repr(i32)]
+pub enum HitOption {
+    NONE,
+    BACK_ATTACK,
+    FRONTAL_ATTACK,
+    FLANK_ATTACK,
+    MAX,
+}
+
+#[derive(Debug, PartialEq)]
+#[allow(non_camel_case_types)]
+#[repr(u32)]
+pub enum HitFlag {
+    NORMAL,
+    CRITICAL,
+    MISS,
+    INVINCIBLE,
+    DOT,
+    IMMUNE,
+    IMMUNE_SILENCED,
+    FONT_SILENCED,
+    DOT_CRITICAL,
+    DODGE,
+    REFLECT,
+    DAMAGE_SHARE,
+    DODGE_HIT,
+    MAX,
+}
