+++ src-tauri/src/models/misc.rs
@@ -1,0 +1,52 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct EncounterPreview {
+    pub id: i32,
+    pub fight_start: i64,
+    pub boss_name: String,
+    pub duration: i64,
+    pub classes: Vec<i32>,
+    pub names: Vec<String>,
+    pub difficulty: Option<String>,
+    pub local_player: String,
+    pub my_dps: i64,
+    pub favorite: bool,
+    pub cleared: bool,
+    pub spec: Option<String>,
+    pub support_ap: Option<f32>,
+    pub support_brand: Option<f32>,
+    pub support_identity: Option<f32>,
+    pub support_hyper: Option<f32>,
+}
+
+#[derive(Debug, Default, Serialize, Deserialize, Clone)]
+#[serde(rename_all = "camelCase")]
+pub struct EncountersOverview {
+    pub encounters: Vec<EncounterPreview>,
+    pub total_encounters: i32,
+}
+
+#[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase", default)]
+pub struct SearchFilter {
+    pub bosses: Vec<String>,
+    pub min_duration: i32,
+    pub max_duration: i32,
+    pub cleared: bool,
+    pub favorite: bool,
+    pub difficulty: String,
+    pub boss_only_damage: bool,
+    pub sort: String,
+    pub order: String,
+    pub raids_only: bool,
+}
+
+#[derive(Default, Debug, Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct EncounterDbInfo {
+    pub size: String,
+    pub total_encounters: i32,
+    pub total_encounters_filtered: i32,
+}
