+++ src-tauri/src/models/status_effect.rs
@@ -1,0 +1,52 @@
+use bitflags::bitflags;
+use serde::{Deserialize, Serialize};
+
+use crate::models::SkillData;
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+#[serde(rename_all = "camelCase")]
+pub struct StatusEffect {
+    pub target: StatusEffectTarget,
+    pub category: String,
+    pub buff_category: String,
+    pub buff_type: u32,
+    pub unique_group: u32,
+    pub source: StatusEffectSource,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize, Default, PartialEq)]
+pub enum StatusEffectTarget {
+    #[default]
+    OTHER,
+    PARTY,
+    SELF,
+}
+
+#[derive(Debug, Clone, Serialize, Default, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct StatusEffectSource {
+    pub name: String,
+    pub desc: String,
+    pub icon: String,
+    pub skill: Option<SkillData>,
+    pub set_name: Option<String>,
+}
+
+bitflags! {
+    #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash)]
+    pub struct StatusEffectBuffTypeFlags: u32 {
+        const NONE = 0;
+        const DMG = 1;
+        const CRIT = 1 << 1;
+        const ATKSPEED = 1 << 2;
+        const MOVESPEED = 1 << 3;
+        const HP = 1 << 4;
+        const DEFENSE = 1 << 5;
+        const RESOURCE = 1 << 6;
+        const COOLDOWN = 1 << 7;
+        const STAGGER = 1 << 8;
+        const SHIELD = 1 << 9;
+
+        const ANY = 1 << 20;
+    }
+}
