+++ src-tauri/src/models/utils.rs
@@ -1,0 +1,14 @@
+use serde::{Deserialize, Deserializer};
+use serde_json::Value;
+
+pub fn int_or_string_as_string<'de, D>(deserializer: D) -> Result<String, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    let value = Value::deserialize(deserializer)?;
+    match value {
+        Value::String(s) => Ok(s),
+        Value::Number(n) => Ok(n.to_string()),
+        _ => Err(serde::de::Error::custom("Expected a string or an integer")),
+    }
+}
