--- src-tauri/src/parser/mod.rs
@@ -1,1 +1,0 @@
-pub mod models;
